requirements.txt
pysocks
requests